feat(add general dark mode): for header, footer \
